+++ libs/components/phone-field/src/app/public/plugin-resources/phone-field-resources-provider.ts
@@ -1,8 +1,0 @@
-import {
-  SkyAppLocaleInfo,
-  SkyLibResourcesProvider
-} from '@skyux/i18n';
-
-export class SkyPhoneFieldResourcesProvider implements SkyLibResourcesProvider {
-  public getString: (localeInfo: SkyAppLocaleInfo, name: string) => string;
-}