+++ libs/components/phone-field/src/app/public/public_api.ts
@@ -1,3 +1,0 @@
-export * from './modules/phone-field/phone-field.module';
-export * from './modules/phone-field/types/country';
-export * from './modules/phone-field/types/number-return-format';